requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.2.9
django-cors-headers==4.9.0
django-ninja==1.5.1
grpcio==1.76.0
grpcio-tools==1.76.0
idna==3.11
protobuf==6.33.2
pydantic==2.12.5
pydantic_core==2.41.5
python-decouple==3.8
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
